LayoutTests/http/tests/notifications/legacy/double-show-expected.txt

Sending double-show notification...

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.webkitNotifications is defined.
PASS ondisplay event handler called
PASS successfullyParsed is true

TEST COMPLETE
